requirements.txt
alembic==1.13.1
asgiref==3.8.1
bcrypt==4.1.2
blinker==1.7.0
cachelib==0.13.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
msgspec==0.18.6
mysqlclient==2.2.4
packaging==24.0
pluggy==1.5.0
pytest==8.2.0
python-dotenv==1.0.1
redis==5.0.3
requests==2.31.0
SQLAlchemy==2.0.28
tenacity==8.2.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
Werkzeug==3.0.1